requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.64.0
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.14
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
